Revet regression from 89e721a
In the commit 89e721a there were two changes. The first was fine, so this commit does not revert that change.

But the second change made in 89e721a was a major regression. It's all about calling Function.prototype.apply supplying fixture.content as the array of arguments. Inspecting the definition of that fixture.contents (line 47) reveals that the property comes from a module called file-reader that returns objects, not arrays. Function.prototype.apply is not appropriate for objects, so this commit ensures that Function.prototype.apply is not used (unless for some reason fixture.content happens to be array-like).